Using Basic Machine for Inheritance
